refactor: decided i dont care about importing prisma in presenters 😘✨
